# payment-service-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Scan Results: payment-service",
    "version": "1.0",
    "description": "APIs discovered from payment-service (./test_samples)",
    "x-generated-by": "Universal Polyglot API Scanner v3.1",
    "x-generated-at": "2026-01-31T23:32:47.898244",
    "x-service-name": "payment-service"
  },
  "paths": {
...
    "/admin": {
      "get": {
        "summary": "Spring endpoint",
        "description": "Discovered from [SpringApp.java]:159\n\nFramework: Spring\nLanguage: Java\nRisk Level: HIGH\nAuth Status: PRIVATE",
        "operationId": "get_admin",
        "tags": [
          "Spring",
          "Java"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\SpringApp.java",
        "x-source-line": 159,
        "x-risk-level": "HIGH",
        "x-auth-status": "PRIVATE",
        "x-risk-reasons": [
          "admin: \\badmin\\b"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
    "/admin/database/reset": {
      "post": {
        "summary": "ASP.NET endpoint",
        "description": "Discovered from [Controllers.cs]:159\n\nFramework: ASP.NET\nLanguage: C#/.NET\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "post_admin_database_reset",
        "tags": [
          "ASP.NET",
          "C#/.NET"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\Controllers.cs",
        "x-source-line": 159,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bauth\\b",
          "admin: \\badmin\\b",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/admin/logs": {
      "get": {
        "summary": "ASP.NET endpoint",
        "description": "Discovered from [Controllers.cs]:173\n\nFramework: ASP.NET\nLanguage: C#/.NET\nRisk Level: HIGH\nAuth Status: UNKNOWN",
        "operationId": "get_admin_logs",
        "tags": [
          "ASP.NET",
          "C#/.NET"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\Controllers.cs",
        "x-source-line": 173,
        "x-risk-level": "HIGH",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "admin: \\badmin\\b",
          "Shadow API"
        ]
      }
    },
...
    "/admin/reset": {
      "get": {
        "summary": "net/http endpoint",
        "description": "Discovered from [go_app.go]:31\n\nFramework: net/http\nLanguage: Go\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "get_admin_reset",
        "tags": [
          "net/http",
          "Go"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\go_app.go",
        "x-source-line": 31,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bauth\\b",
          "admin: \\badmin\\b",
          "Shadow API"
        ]
      }
    },
...
    "/admin/users/{id}": {
      "delete": {
        "summary": "Custom-Router endpoint",
        "description": "Discovered from [complex_python_app.py]:91\n\nFramework: Custom-Router\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "delete_admin_users_id",
        "tags": [
          "Custom-Router",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\complex_python_app.py",
        "x-source-line": 91,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "financial: \\bbilling\\b",
          "admin: \\badmin\\b",
          "admin: \\binternal\\b",
          "Mutation: DELETE",
          "Shadow API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: id"
          }
        ]
      }
    },
    "/admin/users/{user_id}": {
      "delete": {
        "summary": "Flask endpoint",
        "description": "Discovered from [fastapi_app.py]:66\n\nFramework: Flask\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "delete_admin_users_user_id",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\fastapi_app.py",
        "x-source-line": 66,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bauth\\b",
          "admin: \\badmin\\b",
          "Mutation: DELETE",
          "Shadow API"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: user_id"
          }
        ]
      }
    },
...
    "/api/auth/login": {
      "post": {
        "summary": "Express endpoint",
        "description": "Discovered from [express_app.js]:23\n\nFramework: Express\nLanguage: JavaScript\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "post_api_auth_login",
        "tags": [
          "Express",
          "JavaScript"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\express_app.js",
        "x-source-line": 23,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bpassword\\b",
          "auth: \\btoken\\b",
          "auth: \\bauth\\b",
          "auth: \\bjwt\\b",
          "pii: \\bemail\\b",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Flask endpoint",
        "description": "Discovered from [flask_app.py]:28\n\nFramework: Flask\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "get_api_auth_login",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\flask_app.py",
        "x-source-line": 28,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bpassword\\b",
          "auth: \\bauth\\b",
          "pii: \\bemail\\b",
          "Shadow API"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Express endpoint",
        "description": "Discovered from [express_app.js]:29\n\nFramework: Express\nLanguage: JavaScript\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "post_api_auth_register",
        "tags": [
          "Express",
          "JavaScript"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\express_app.js",
        "x-source-line": 29,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bpassword\\b",
          "auth: \\btoken\\b",
          "auth: \\bauth\\b",
          "pii: \\bssn\\b",
          "pii: \\bemail\\b",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Flask endpoint",
        "description": "Discovered from [flask_app.py]:35\n\nFramework: Flask\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "get_api_auth_register",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\flask_app.py",
        "x-source-line": 35,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bpassword\\b",
          "auth: \\btoken\\b",
          "auth: \\bauth\\b",
          "pii: \\bssn\\b",
          "Shadow API"
        ]
      }
    },
...
    "/api/orders": {
      "get": {
        "summary": "Fiber endpoint",
        "description": "Discovered from [go_app.go]:176\n\nFramework: Fiber\nLanguage: Go\nRisk Level: LOW\nAuth Status: UNKNOWN",
        "operationId": "get_api_orders",
        "tags": [
          "Fiber",
          "Go"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\go_app.go",
        "x-source-line": 176,
        "x-risk-level": "LOW",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "Shadow API"
        ]
      },
      "post": {
        "summary": "Fiber endpoint",
        "description": "Discovered from [go_app.go]:178\n\nFramework: Fiber\nLanguage: Go\nRisk Level: LOW\nAuth Status: UNKNOWN",
        "operationId": "post_api_orders",
        "tags": [
          "Fiber",
          "Go"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\go_app.go",
        "x-source-line": 178,
        "x-risk-level": "LOW",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "summary": "Fiber endpoint",
        "description": "Discovered from [go_app.go]:177\n\nFramework: Fiber\nLanguage: Go\nRisk Level: LOW\nAuth Status: UNKNOWN",
        "operationId": "get_api_orders_id",
        "tags": [
          "Fiber",
          "Go"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\go_app.go",
        "x-source-line": 177,
        "x-risk-level": "LOW",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "Shadow API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: id"
          }
        ]
      },
      "put": {
        "summary": "Fiber endpoint",
        "description": "Discovered from [go_app.go]:179\n\nFramework: Fiber\nLanguage: Go\nRisk Level: LOW\nAuth Status: UNKNOWN",
        "operationId": "put_api_orders_id",
        "tags": [
          "Fiber",
          "Go"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\go_app.go",
        "x-source-line": 179,
        "x-risk-level": "LOW",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "Mutation: PUT",
          "Shadow API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: id"
          }
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Fiber endpoint",
        "description": "Discovered from [go_app.go]:180\n\nFramework: Fiber\nLanguage: Go\nRisk Level: LOW\nAuth Status: UNKNOWN",
        "operationId": "delete_api_orders_id",
        "tags": [
          "Fiber",
          "Go"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\go_app.go",
        "x-source-line": 180,
        "x-risk-level": "LOW",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "Mutation: DELETE",
          "Shadow API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter: id"
          }
        ]
      }
    },
    "/api/payments/charge": {
      "post": {
        "summary": "Flask endpoint",
        "description": "Discovered from [fastapi_app.py]:55\n\nFramework: Flask\nLanguage: Python\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "post_api_payments_charge",
        "tags": [
          "Flask",
          "Python"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\fastapi_app.py",
        "x-source-line": 55,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\btoken\\b",
          "pii: \\bemail\\b",
          "financial: \\bpayment\\b",
          "financial: \\bcharge\\b",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/products": {
      "get": {
        "summary": "Echo endpoint",
        "description": "Discovered from [go_app.go]:150\n\nFramework: Echo\nLanguage: Go\nRisk Level: LOW\nAuth Status: UNKNOWN",
        "operationId": "get_api_products",
        "tags": [
          "Echo",
          "Go"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\go_app.go",
        "x-source-line": 150,
        "x-risk-level": "LOW",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "Shadow API"
        ]
      },
      "post": {
        "summary": "Echo endpoint",
        "description": "Discovered from [go_app.go]:152\n\nFramework: Echo\nLanguage: Go\nRisk Level: LOW\nAuth Status: UNKNOWN",
        "operationId": "post_api_products",
        "tags": [
          "Echo",
          "Go"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\go_app.go",
        "x-source-line": 152,
        "x-risk-level": "LOW",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/search/products": {
      "get": {
        "summary": "Express endpoint",
        "description": "Discovered from [express_app.js]:73\n\nFramework: Express\nLanguage: JavaScript\nRisk Level: LOW\nAuth Status: UNKNOWN",
        "operationId": "get_api_search_products",
        "tags": [
          "Express",
          "JavaScript"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\express_app.js",
        "x-source-line": 73,
        "x-risk-level": "LOW",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "Shadow API"
        ]
      }
    },
...
    "/api/users/": {
      "get": {
        "summary": "net/http endpoint",
        "description": "Discovered from [go_app.go]:28\n\nFramework: net/http\nLanguage: Go\nRisk Level: CRITICAL\nAuth Status: UNKNOWN",
        "operationId": "get_api_users",
        "tags": [
          "net/http",
          "Go"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\go_app.go",
        "x-source-line": 28,
        "x-risk-level": "CRITICAL",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "auth: \\bauth\\b",
          "admin: \\badmin\\b",
          "Shadow API"
        ]
      }
    },
...
    "/api/users/search": {
      "get": {
        "summary": "ASP.NET endpoint",
        "description": "Discovered from [Controllers.cs]:33\n\nFramework: ASP.NET\nLanguage: C#/.NET\nRisk Level: LOW\nAuth Status: UNKNOWN",
        "operationId": "get_api_users_search",
        "tags": [
          "ASP.NET",
          "C#/.NET"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "test_samples\\Controllers.cs",
        "x-source-line": 33,
        "x-risk-level": "LOW",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "Shadow API"
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}